backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
sqlalchemy==2.0.35
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.5.2
statsmodels==0.14.4
textblob==0.18.0
fpdf2==2.8.1
reportlab==4.2.5
pydantic==2.9.2
python-multipart==0.0.12
websockets==13.1
pymongo[srv]==4.9.1
bcrypt==4.2.0
PyJWT==2.9.0
python-dotenv==1.0.1
google-genai==1.1.0
dnspython==2.7.0
psutil==6.0.0
langgraph>=0.2.0
langchain-core>=0.3.0
